# swagger/swagger.yaml
...
swagger: "2.0"
info:
  version: "1.0.0"
  title: "Hyperledger Fabric Gateway API"
host: "localhost:4000"
basePath: "/"
schemes:
- "http"
produces:
- "application/json"
paths:
...
  /monitor/peer/info:
    get:
      tags:
      - "Monitor"
      summary: "Get peer's info"
      description: "Get peer's info"
      operationId: "returnPeerInfo"
      produces:
      - "application/json"
      parameters:
      - name: "Authorization"
        in: "header"
        description:  "asccess token"
        required: true
        type: "string"
      - name: "peerName"
        in: "query"
        description: "peerName you want to query info"
        required: true
        # example: peer1
        type: "string"
      responses:
        200:
          description: "SDK response"
          schema:
            $ref: "#/definitions/sdkResoponse"
        default:
          description: "Unexpected error"
          schema:
            $ref: "#/definitions/inline_response_default"
      x-swagger-router-controller: "monitor"
...
  /monitor/channel/info:
    get:
      tags:
      - "Monitor"
      summary: "Get channel's info"
      description: "Get channel's info"
      operationId: "returnChannelInfo"
      produces:
      - "application/json"
      parameters:
      - name: "Authorization"
        in: "header"
        description:  "asccess token"
        required: true
        type: "string"
      - name: "peerName"
        # example: peer1
        in: "query"
        description: "peerName you want to query info"
        required: true
        type: "string"
      - name: "channelName"
        in: "query"
        # example: mychannel
        description: "channelName you want to query info"
        required: true
        type: "string"
      responses:
        200:
          description: "SDK response"
          schema:
            $ref: "#/definitions/sdkResoponse"
        default:
          description: "Unexpected error"
          schema:
            $ref: "#/definitions/inline_response_default"
      x-swagger-router-controller: "monitor"
  /monitor/block/info:
    get:
      tags:
      - "Monitor"
      summary: "Get block's info"
      description: "Get block's info"
      operationId: "returnBlockInfo"
      produces:
      - "application/json"
      parameters:
      - name: "Authorization"
        in: "header"
        description:  "asccess token"
        required: true
        type: "string"
      - name: "peerName"
        # example: peer1
        in: "query"
        description: "peerName you want to query info"
        required: true
        type: "string"
      - name: "channelName"
        in: "query"
        # example: mychannel
        description: "channelName you want to query info"
        required: true
        type: "string"
      - name: "blockNumber"
        in: "query"
        # example: "1"
        description: "blockNumber you want to query info"
        required: true
        type: "string"
      responses:
        200:
          description: "SDK response"
          schema:
            $ref: "#/definitions/sdkResoponse"
        default:
          description: "Unexpected error"
          schema:
            $ref: "#/definitions/inline_response_default"
      x-swagger-router-controller: "monitor"
  /monitor/tx/last:
    get:
      tags:
      - "Monitor"
      summary: "Get lastest tx info"
      description: "Get tx's info"
      operationId: "returnLastTxs"
      produces:
      - "application/json"
      parameters:
      - name: "Authorization"
        in: "header"
        description:  "asccess token"
        required: true
        type: "string"
      - name: "peerName"
        in: "query"
        # example: peer1
        description: "peerName you want to query info"
        required: true
        type: "string"
      - name: "channelName"
        in: "query"
        # example: mychannel
        description: "channelName you want to query info"
        required: true
        type: "string"
      - name: "lastNumber"
        in: "query"
        # example: "20"
        description: "How many tx want to query info"
        required: true
        type: "string"
      responses:
        200:
          description: "SDK response"
          schema:
            $ref: "#/definitions/sdkResoponse"
        default:
          description: "Unexpected error"
          schema:
            $ref: "#/definitions/inline_response_default"
      x-swagger-router-controller: "monitor"
  /monitor/chaincode/info:
    get:
      tags:
      - "Monitor"
      summary: "Get chaincode info"
      description: "Get chaincode info"
      operationId: "returnChaincodeInfo"
      produces:
      - "application/json"
      parameters:
      - name: "Authorization"
        in: "header"
        description:  "asccess token"
        required: true
        type: "string"
      - name: "peerName"
        in: "query"

          # example: peer1
        description: "peerName you want to query info"
        required: true
        type: "string"
      - name: "channelName"
        in: "query"
        # example: mychannel
        description: "channelName you want to query info"
        required: true
        type: "string"
      - name: "chaincodeName"
        in: "query"
        # example: mycc
        description: "chaincode name you want to query info"
        required: true
        type: "string"
      responses:
        200:
          description: "SDK response"
          schema:
            $ref: "#/definitions/sdkResoponse"
        default:
          description: "Unexpected error"
          schema:
            $ref: "#/definitions/inline_response_default"
      x-swagger-router-controller: "monitor"
...
  /monitor/network/my:
    get:
      tags:
      - "Monitor"
      summary: "Get self network-config"
      description: "Get self network-config"
      operationId: "returnSelfNetworkConfig"
      produces:
      - "application/json"
      parameters:
      - name: "Authorization"
        in: "header"
        description:  "asccess token"
        required: true
        type: "string"
      responses:
        200:
          description: "SDK response"
          schema:
            $ref: "#/definitions/sdkResoponse"
        default:
          description: "Unexpected error"
          schema:
            $ref: "#/definitions/inline_response_default"
      x-swagger-router-controller: "monitor"
...
definitions:
...
  inline_response_default:
    properties:
      code:
        type: "integer"
        format: "int32"
      message:
        type: "string"
      fields:
        type: "string"
...
  sdkResoponse:
    properties:
      sdkResult:
        type: "object"
...
securityDefinitions:
   # X-API-Key: abcdef12345
   APIKeyHeader:
     type: apiKey
     in: header
     name: X-API-Key